Health and Damage to Enemy
The enemy cannot take Damage to their health yet as I'm having a bit of a problem with their collision and such, but as soon as that works they should be able to be destroyed and take damage. So for now I'm pushing this just so everyone has it, and there won't be any conflicts in the future.